some changes in admin card

+++ Server/controllers/Reward.js
@@ -3,45 +3,45 @@ import { respond } from "../utils/response.js"
 import { Card } from "../models/Card.js";
 import mongoose from "mongoose";
 
-// export const addRewards = async (req, res) => {
-//     try {
-//         const { instruction, points, listData, note,cardId } = req.body;
+export const addRewards = async (req, res) => {
+    try {
+        const { instruction, points, listData, note,cardId } = req.body;
 
-//         // Validate input
-//         // if (points && !Array.isArray(points)) {
-//         //     return respond(res, "Points should be an array", 400, false);
-//         // }
-//         // if (listData && !Array.isArray(listData)) {
-//         //     return respond(res, "List data should be an array", 400, false);
-//         // }
+        // Validate input
+        // if (points && !Array.isArray(points)) {
+        //     return respond(res, "Points should be an array", 400, false);
+        // }
+        // if (listData && !Array.isArray(listData)) {
+        //     return respond(res, "List data should be an array", 400, false);
+        // }
 
-//         // Create a new Reward document
-//         const reward = new Reward({
-//             instruction,
-//             points: points.length > 0 ? points.map(point => ({
-//                 _id: new mongoose.Types.ObjectId(),
-//                 key: point.key,
-//                 value: point.value
-//             })) : [],
-//             listData,
-//             note
-//         });
+        // Create a new Reward document
+        const reward = new Reward({
+            instruction,
+            points: points.length > 0 ? points.map(point => ({
+                _id: new mongoose.Types.ObjectId(),
+                key: point.key,
+                value: point.value
+            })) : [],
+            listData,
+            note
+        });
 
-//         // Save the Reward document
-//         const savedReward = await reward.save();
+        // Save the Reward document
+        const savedReward = await reward.save();
 
-//         const updatedCard = await Card.findByIdAndUpdate(
-//             cardId,
-//             { $push: {rewards : savedReward._id } },
-//             { new: true }
-//         ).populate("rewards");
+        const updatedCard = await Card.findByIdAndUpdate(
+            cardId,
+            { $push: {rewards : savedReward._id } },
+            { new: true }
+        ).populate("rewards");
 
-//         return respond(res, "Rewards added successfully", 200, true, savedReward);
-//     } catch (error) {
-//         console.log(error);
-//         return respond(res, "Something went wrong while adding the rewards", 500, false);
-//     }
-// };
+        return respond(res, "Rewards added successfully", 200, true, savedReward);
+    } catch (error) {
+        console.log(error);
+        return respond(res, "Something went wrong while adding the rewards", 500, false);
+    }
+};
 
 // export const addRewards = async (req, res) => {
 //     try {
@@ -83,58 +83,61 @@ import mongoose from "mongoose";
 //     }
 // };
 
-export const addRewards = async (req, res) => {
-    try {
-        const { instruction, note, cardId } = req.body;
+// Add Rewards
+// export const addRewards = async (req, res) => {
+//     try {
+//         const { instruction, note, cardId } = req.body;
 
-        // Manually build the points array
-        const points = [];
-        if (req.body["points[0][key]"]) {
-            let i = 0;
-            while (req.body[`points[${i}][key]`]) {
-                points.push({
-                    key: req.body[`points[${i}][key]`],
-                    value: req.body[`points[${i}][value]`],
-                });
-                i++;
-            }
-        }
+//         // Manually build the points array
+//         const points = [];
+//         if (req.body["points[0][key]"]) {
+//             let i = 0;
+//             while (req.body[`points[${i}][key]`]) {
+//                 points.push({
+//                     _id: new mongoose.Types.ObjectId(), // Generate unique ID for each point
+//                     key: req.body[`points[${i}][key]`],
+//                     value: req.body[`points[${i}][value]`],
+//                 });
+//                 i++;
+//             }
+//         }
 
-        // Manually build the listData array
-        const listData = [];
-        if (req.body["listData[0]"]) {
-            let i = 0;
-            while (req.body[`listData[${i}]`]) {
-                listData.push(req.body[`listData[${i}]`]);
-                i++;
-            }
-        }
+//         // Manually build the listData array
+//         const listData = [];
+//         if (req.body["listData[0]"]) {
+//             let i = 0;
+//             while (req.body[`listData[${i}]`]) {
+//                 listData.push(req.body[`listData[${i}]`]);
+//                 i++;
+//             }
+//         }
 
-        // Create a new Reward document
-        const reward = new Reward({
-            instruction,
-            points,
-            listData,
-            note,
-        });
+//         // Create a new Reward document
+//         const reward = new Reward({
+//             instruction,
+//             points,
+//             listData,
+//             note,
+//         });
 
-        // Save the Reward document
-        const savedReward = await reward.save();
+//         // Save the Reward document
+//         const savedReward = await reward.save();
 
-        // Update the corresponding card with the new reward ID
-        const updatedCard = await Card.findByIdAndUpdate(
-            cardId,
-            { $push: { rewards: savedReward._id } },
-            { new: true }
-        ).populate("rewards");
+//         // Update the corresponding card with the new reward ID
+//         const updatedCard = await Card.findByIdAndUpdate(
+//             cardId,
+//             { $push: { rewards: savedReward._id } },
+//             { new: true }
+//         ).populate("rewards");
 
-        return respond(res, "Rewards added successfully", 200, true, savedReward);
-    } catch (error) {
-        console.log(error);
-        return respond(res, "Something went wrong while adding the rewards", 500, false);
-    }
-};
+//         return respond(res, "Rewards added successfully", 200, true, updatedCard);
+//     } catch (error) {
+//         console.log(error);
+//         return respond(res, "Something went wrong while adding the rewards", 500, false);
+//     }
+// };
 
+// Update Rewards
 export const updateRewards = async (req, res) => {
     try {
         const { rewardId, instruction, points, listData, note } = req.body;
@@ -159,8 +162,8 @@ export const updateRewards = async (req, res) => {
             reward.note = note;
         }
 
-        if (listData !== undefined) {
-            reward.listData = listData;
+        if (Array.isArray(listData)) {
+            reward.listData = listData; // Replace existing listData
         }
 
         // Update points if provided
@@ -197,25 +200,32 @@ export const updateRewards = async (req, res) => {
     }
 };
 
+// Delete Rewards
+export const deleteRewards = async (req, res) => {
+    try {
+        const { rewardId, cardId } = req.body;
 
-export const deleteRewards = async(req,res) => {
-    try{
-        const {rewardId} = req.body
-
-        if(!rewardId) {
-            return respond(res,"reward id is not found",400,false)
+        if (!rewardId) {
+            return respond(res, "Reward ID is required", 400, false);
         }
 
-        const updetedCard = await Card.findByIdAndUpdate(
-            {rewards:rewardId},
-            {$pull: {rewards:rewardId}}
-        ).populate("rewardId")
+        // Remove the reward reference from the Card
+        const updatedCard = await Card.findByIdAndUpdate(
+            cardId,
+            { $pull: { rewards: rewardId } },
+            { new: true }
+        ).populate("rewards");
 
-        await Reward.findByIdAndDelete(rewardId)
+        if (!updatedCard) {
+            return respond(res, "Card not found", 404, false);
+        }
 
-        return respond(res,"Rewards deleted successfully",200,true,updetedCard)
-    }catch(error) {
-        console.log(error)
-        return respond(res,"something went wrong while deleting the rewards",500,false)
+        // Delete the Reward document
+        await Reward.findByIdAndDelete(rewardId);
+
+        return respond(res, "Rewards deleted successfully", 200, true, updatedCard);
+    } catch (error) {
+        console.log(error);
+        return respond(res, "Something went wrong while deleting the rewards", 500, false);
     }
-}
+};
